Digital filter design demo.
First cut at implementing a digital filter design page.

Currently only supports Butterworth filters.  The user enters the pass
band and stop band frequencies, and the desired attenutation in the pass
band and stop band.

The resulting Butterworth filter is designed, with the tranform
displayed.  The equivalent digital filter using a bilinear transform
is displayed.

Finally, code is displayed that implements the digital filter as a
sequence of biquad filters with possibly one IIR filter.

Needs verification.  And if this is correct, it clearly shows that the
new biquad filter equations are needed to be able to implement these
filters.