Inline just the css we need from examples.css
examples.css did a lot of things that we don't really need.  So, just
get the parts we need to style the graph and place it inline.

The most important part is css for the legend so that we can actually
read the text.  Without this the legend was a black box hiding the
text labels.