Make sliders line up with axis
The purpose of this is to make the cutoff slider line up with the
graph so the the slider value matches the point on the graph.  For
example, the slider should be under the peak of the bandpass filter.

The location of the other sliders isn't relevant, but we make them all
the same size and location.

This isn't perfect.  The slider doesn't quite match up with the axis,
but close enough for now.

Also added support for changing these when the window is resized so
the slider matches graph.

Also, the values in the cutoff slider were not quite right.  Rearrange
so that slider values are logarithmic between log10(lowest) to
log10(Nyquist).